Rename transformer_pt.py to transformer.py
